implement backwards compatible pairwise MACs for exploding messages
For exploding messages on small teams, include pairwise MACs in the
header. Clients with no pairwise MAC support will ignore these, and
we'll keep signing with the regular key, so there should be no breakage.
Define message boxed V4, which will be when we stop signing with the
regular key, but don't emit those messages yet.